fix problem with spice config values not reading

+++ src/main/java/schmille/cabinetofmadness/properties/SpiceProperties.java
@@ -7,8 +7,8 @@ public class SpiceProperties {
     private final ForgeConfigSpec.IntValue time;
     private final ForgeConfigSpec.IntValue nutrition;
     private final ForgeConfigSpec.IntValue effectLevel;
-    private final ForgeConfigSpec.ConfigValue<Float> saturation;
-    private final ForgeConfigSpec.ConfigValue<Float> probability;
+    private final ForgeConfigSpec.ConfigValue<Double> saturation;
+    private final ForgeConfigSpec.ConfigValue<Double> probability;
 
     SpiceProperties(ForgeConfigSpec.Builder builder) {
         builder.push("Spice");
@@ -16,8 +16,8 @@ public class SpiceProperties {
                 .comment("time in seconds")
                 .defineInRange("spice_effect_time", 25,1, 24 * 60 * 60);
         nutrition = builder.defineInRange("spice_nutrition", 2, 1, 16);
-        saturation = builder.define("spice_saturation", 1.0F);
-        probability = builder.define("spice_effect_probability", 1.0F);
+        saturation = builder.define("spice_saturation", 1.0D);
+        probability = builder.define("spice_effect_probability", 1.0D);
         effectLevel = builder
                 .comment("effect level starting at 1")
                 .defineInRange("spice_effect_level",1, 1, 4);
@@ -33,11 +33,11 @@ public class SpiceProperties {
     }
 
     public float saturation() {
-        return saturation.get();
+        return saturation.get().floatValue();
     }
 
     public float probability() {
-        return probability.get();
+        return probability.get().floatValue();
     }
 
     public int level() {
